Ignore unmanaged files
Sometimes, you have a file that needs to be specific to some nodes/clusters, in ways that would make no sense to really manage centrally.

For these, you can blacklist them! Just create a file called '_ignore' in the configuration directory and fill it with the names of ignored files. One file per line.